Add tickless idle, port guards, and various fixes
Centralize and finish tickless-idle support and related correctness fixes. Introduce arch/common/os_arch_tickless.c (Cortex-M adapter in arch/arm/common) and require OS_ARCH_PORT_TRANSLATION_UNIT for textual includes; wrapper port files now define that macro. Harden tickless behaviour (ceilings/floors, clamp, planning guards, cross-core nudge) and fix fractional-tick accumulation in SoC drivers to avoid clock drift. Add link-anchor check in CMake and force soc package anchors to avoid silent archive drops. Add data-barrier (DMB) macros and use them for atomic ordering, add sleep mask swap helpers, refine OS_ARCH_SLEEP/PendSV behaviour. Improve task/timer/delay delta lists and task id generation, make mutex inheritance transitive with chain recompute, and expand self-test (SVC vector option, tickless suppression test). Update docs and examples accordingly.

+++ kernel/os_mem.c
@@ -345,8 +345,17 @@ static void os_mem_block_insert(os_mem_block_t *block)
     os_mem_block_t *iter;
     uint8_t        *address;
 
-    /* Find the free block we insert after (list is address ordered). */
-    for (iter = &os_mem_start; iter->next < block; iter = iter->next)
+    /* Find the free block we insert after (list is address ordered).
+     *
+     * Compared as integers rather than as pointers, and that is not pedantry dressed up: the walk
+     * STARTS at os_mem_start, which is a static living outside os_mem_heap[] entirely. Relational
+     * comparison between pointers into different objects is undefined in C - it happens to work on
+     * every flat address space, which is exactly what makes it the kind of thing a compiler is free
+     * to surprise you with later, and what a MISRA (Rule 18.3) or CERT run flags. Converting both
+     * sides costs nothing and the question being asked is genuinely about addresses. */
+    for (iter = &os_mem_start;
+         (uintptr_t)iter->next < (uintptr_t)block;
+         iter = iter->next)
     {
     }
